Tidy up modules
