+++ lib/Cargo.toml
@@ -11,3 +11,5 @@ snow = "0.8"
 tracing = "0.1"
 packet = "0.1"
 base64 = "0"
+regex = "1"
+lazy_static = "1.4"
