+++ lib/src/lib.rs
@@ -1,3 +1,4 @@
 mod errors;
 
 pub mod gateway;
+pub mod identity;
